Python Automation QA Sample/requirements.txt
allure-behave==2.13.5
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.4.0
behave==1.2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
colorlog==6.8.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
platformdirs==4.3.2
pluggy==1.6.0
pycparser==2.23
PySocks==1.7.1
pytest==8.2.2
pytest-bdd==7.1.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.3
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
